wasmparser: move component-related core sections.
This commit moves the component-related core section parsing out of the `core`
module and into the `component` module in `wasmparser`.

Originally these were put in the core module in anticipation of one day having
the sections be part of core, but that's a long way off and may not occur.
Until that time, it makes sense to keep these sections grouped with the other
component sections.

This commit also removes the extensions to the `ExternalKind` enumeration in
core.

It updates alias parsing to limit outer alias kind to only the supported kinds
(core module, types, and components). This helps reduce confusion between
what's implemented for export aliasing and what's supported for outer aliasing.

By doing so, this also enables the ability to reference outer core types in
nested components.