wast: fix parsing and encoding of aliases.
This commit fixes the parsing and encoding of aliases such that the grammar is
restricted to what's actually supported for both instance export aliases and
for outer aliases.

It also addresses the inability to reference and encode outer core type
aliases.

`CoreItemRef` was fixed to only have a single export name supported as core
instances cannot export other core instances, so there can be at most one
string parsed for the export name.